Make request_timeout_seconds a whole number of seconds
qdrant-client takes its timeout as an int and rounds a fraction up, so a
fractional value was honored by pymilvus and silently changed for Qdrant.
An int is honored exactly by both, and matches max_retry_interval_seconds
on the embedder and language model configurations.

Folded from #1630 into #1631: the whole seconds reach the configurations and
tests that also name a collection_registry.

(cherry picked from commit 5e9f36385b39ff645830ec60f4af4b6f544e2736)
(cherry picked from commit 874ec470738810654068c845a8b5642716850ae3)

+++ packages/server/src/memmachine_server/common/configuration/database_conf.py
@@ -274,8 +274,8 @@ class QdrantConf(MetricsFactoryIdMixin, YamlSerializableMixin, ApiKeyMixin):
             "above the longest a request to Qdrant can be in flight."
         ),
     )
-    request_timeout_seconds: float = Field(
-        default=30.0,
+    request_timeout_seconds: int = Field(
+        default=30,
         gt=0,
         description="Seconds a request to Qdrant may take before the client gives up.",
     )
@@ -331,8 +331,8 @@ class MilvusConf(YamlSerializableMixin, WithValueFromEnv):
             "above the longest a request to Milvus can be in flight."
         ),
     )
-    request_timeout_seconds: float = Field(
-        default=30.0,
+    request_timeout_seconds: int = Field(
+        default=30,
         gt=0,
         description="Seconds a request to Milvus may take before the client gives up.",
     )
